requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-tinymce==3.2.0
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
parso==0.7.1
Pillow==7.2.0
pluggy==0.13.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==2.0.0
pylint==2.6.0
pyls==0.1.6
python-dotenv==0.15.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-social-auth==0.2.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==0.3.1
rope==0.18.0
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.2
ujson==4.0.1
urllib3==1.26.2
wrapt==1.12.1
yapf==0.30.0
